AI/requirements.txt
fastapi==0.95.1
uvicorn==0.23.2
sqlalchemy==2.0.30
python-dotenv==1.0.0
langchain==0.2.1
langchain_community==0.2.1
qdrant_client==1.9.1
openai==1.30.3
langchain-openai==0.1.7
unstructured==0.14.2
boto3==1.34.113
unstructured[pdf]